requirements.txt
google-generativeai==0.8.4
streamlit==1.41.0
pypdf==5.1.0
chromadb==0.6.3
duckduckgo-search==7.1.0
python-dotenv==1.0.1
tiktoken==0.8.0
requests==2.32.3
beautifulsoup4==4.12.3
crewai>=0.80.0
fpdf2==2.8.2
